main_project/django-container/requirements.txt
Django>=4.0
psycopg2-binary
psycopg2==2.9.3
djangorestframework==3.13.1
